Added draftmode, refactor ucfgfx, and autoref fix
You can now use `make draft` in order ot quickly build a draft version
that doesnt include any graphics. Hyperref will still work in non draft
mode, which is desired because we want to be able to jump around in the
document.

The \ucfgfx macro got changed to have an optional argument for changing
the \includegraphics sizing argument. Also a generic \ucffig was created
for later generic figure usage.